fixup! Fix bug when not all alices send read-to-sign
